+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Check if order column in browsers local storage is valid before applying it (report on Gitter).
- 
+- [JENKINS-55337](https://issues.jenkins-ci.org/browse/JENKINS-55337): Show Maven warnings in console log view
 
 ## [1.0.1](https://github.com/jenkinsci/warnings-ng-plugin/compare/warnings-ng-1.0.0...warnings-ng-1.0.1) - 2018-12-28